realtime share improvement
user now allow to input decimal point and RS to individual group child

+++ app/src/main/java/com/eevdf/scheduler/model/Task.kt
@@ -71,7 +71,7 @@ data class Task(
     var startTimeEpoch: Long = 0L,
 
     // CPU share pinning — null = auto-float (EEVDF weight-based), 0–100 = fixed %
-    val pinnedShare: Int? = null,
+    val pinnedShare: Double? = null,
 
     /**
      * Auto-calculated internal scheduling weight derived from [pinnedShare].
